extract documents, disclaimers + storage
