+++ pkg/documents/testdata/colorado.jpg
